+++ test.sh
@@ -18,4 +18,11 @@ rm -f feed.xml
 grep '<h3 id="a-possible-llama-under-the-bridge">A Possible Llama Under the Bridge</h3>' feed.html > /dev/null || exit 1
 rm -f feed.html
 
+# 5----
+./bin/feedmark "eg/Ancient Llama Sightings.md" --output-json >feed.json || exit 1
+grep '"title": "Ancient Llama Sightings"' feed.json > /dev/null || exit 1
+grep '"title": "Maybe sighting the llama"' feed.json > /dev/null || exit 1
+grep '"date": "Jan 1 1984 12:00:00"' feed.json > /dev/null || exit 1
+rm -f feed.json
+
 echo "All tests pass."